Migrate play server to explicit interpreters
