+++ server/play-server/src/main/scala/sttp/tapir/server/play/OutputToPlayResponse.scala
@@ -11,8 +11,6 @@ import akka.util.ByteString
 import play.api.http.{ContentTypes, HeaderNames, HttpEntity}
 import play.api.mvc.MultipartFormData.{DataPart, FilePart}
 import play.api.mvc.{Codec, MultipartFormData, ResponseHeader, Result}
-import sttp.capabilities
-import sttp.capabilities.Streams
 import sttp.model.{MediaType, Part, StatusCode}
 import sttp.tapir.internal.{NoStreams, ParamsAsAny}
 import sttp.tapir.server.internal.{EncodeOutputBody, EncodeOutputs, OutputValues}
